Replace HTTP 304 with HTTP 503 and an error message in index.php
Prior to this commit, when the `topics` table is empty or the connection or query to the database failed due to something, the front page index.php returns the HTTP 304 Not Modified header and no entity body. This has been found faulty in the case when a user hasn't had the page cached by his browser yet, so the browser would simply do nothing. After this commit, in such cases the front page index.php return HTTP 503 Service Unavailable and a proper entity body with a user-friendly HTML error message informing the user about the matter.